+++ hs-bindgen/src-internal/HsBindgen/Frontend/Pass.hs
@@ -134,10 +134,6 @@ class (
   -- For single-constructor datatypes, the index can simply be the name of the
   -- datatype; for multi-constructor datatypes, it should be @"Type.Constr"@
   -- (or simply @"Constr"@ if that is unambiguous).
-  --
-  -- TODO: We could consider using a closed universe for these indices; that
-  -- would make certain things a bit easier (we should show that all annotations
-  -- are showable, for example).
   type Ann (ix :: Symbol) p :: Star
 
   -- | Trace messages possibly emitted by the pass